+++ .gitignore
@@ -1,9 +1,6 @@
 data/
 tempCodeRunnerFile.py
-seg/
-cls/
 __pycache__/
 .DS_Store
 attention_sample.py
-basic.py
-basic2.py
+*.pt
