--- main.sh
@@ -1,1 +1,0 @@
-python main.py --epoch 200  --radius 1 --step_size 50 --points_taken 4096 --batch_size 32 --grid_size 25 --model_name 3 --model pointnet
