+++ pointnet/__init__.py
@@ -1,0 +1,2 @@
+from .pointnet import PointnetSeg
+from .pointnet import PointnetCls
