+++ pointnet2/__init__.py
@@ -1,0 +1,3 @@
+from .utils import farthest_point_sample
+from .utils import index_points
+from .pointnet2 import Pointnet2Seg
